+++ lib/ig_markets/streaming/working_order_update.rb
@@ -1,0 +1,23 @@
+module IGMarkets
+  module Streaming
+    # Contains details on a working order update received via the streaming API. Used by
+    # {DealingPlatform::StreamingMethods#pop_data}.
+    class WorkingOrderUpdate < Model
+      attribute :account_id
+      attribute :channel
+      attribute :deal_id
+      attribute :deal_reference
+      attribute :deal_status, Symbol, allowed_values: [:accepted, :rejected]
+      attribute :direction, Symbol, allowed_values: [:buy, :sell]
+      attribute :epic, String, regex: Regex::EPIC
+      attribute :expiry, Date, nil_if: %w(- DFB), format: ['%d-%b-%y', '%b-%y']
+      attribute :guaranteed_stop, Boolean
+      attribute :level, Float
+      attribute :limit_distance, Fixnum
+      attribute :size, Float
+      attribute :status, Symbol, allowed_values: [:deleted, :open, :updated]
+      attribute :stop_distance, Fixnum
+      attribute :timestamp, Time, format: '%FT%T.%L'
+    end
+  end
+end