+++ spec/ig_markets/model_spec.rb
@@ -108,7 +108,7 @@ describe IGMarkets::Model do
     model.float = 1.0
     model.float = ''
 
-    expect(model.float).to be_nil
+    expect(model.float).to be nil
   end
 
   it 'raises on an invalid boolean' do
@@ -142,7 +142,7 @@ describe IGMarkets::Model do
 
   it 'sets attribute to nil when value matches a nil_if' do
     model.string = '-'
-    expect(model.string).to be_nil
+    expect(model.string).to be nil
   end
 
   it 'parses a date in the expected formats' do